+++ src/js/validator/lessThan.js
@@ -1,13 +1,7 @@
 (function($) {
     $.fn.bootstrapValidator.i18n.lessThan = $.extend($.fn.bootstrapValidator.i18n.lessThan || {}, {
         'default': 'Please enter a value less than or equal to %s',
-        notInclusive: 'Please enter a value less than %s',
-
-        getMessage: function(options) {
-            return (options.inclusive === true || options.inclusive === undefined)
-                    ? $.fn.bootstrapValidator.helpers.format(this['default'], options.value)
-                    : $.fn.bootstrapValidator.helpers.format(this.notInclusive, options.value);
-        }
+        notInclusive: 'Please enter a value less than %s'
     });
 
     $.fn.bootstrapValidator.validators.lessThan = {
@@ -34,18 +28,38 @@
          * @param {BootstrapValidator} validator The validator plugin instance
          * @param {jQuery} $field Field element
          * @param {Object} options Can consist of the following keys:
-         * - value: The number used to compare to
+         * - value: The number used to compare to. It can be
+         *      - A number
+         *      - Name of field which its value defines the number
+         *      - Name of callback function that returns the number
+         *      - A callback function that returns the number
+         *
          * - inclusive [optional]: Can be true or false. Default is true
          * - message: The invalid message
-         * @returns {Boolean}
+         * @returns {Object}
          */
         validate: function(validator, $field, options) {
             var value = $field.val();
             if (value === '') {
                 return true;
             }
+
+            var compareTo = options.value;
+            if ('function' === typeof compareTo) {
+                compareTo = $.fn.bootstrapValidator.helpers.call(compareTo, [value, validator, $field]);
+            } else if ('string' === typeof compareTo && !$.isNumeric(compareTo)) {
+                var $compareField = validator.getFieldElements(compareTo);
+                if ($compareField.length) {
+                    compareTo = $compareField.val();
+                } else {
+                    compareTo = $.fn.bootstrapValidator.helpers.call(compareTo, [value, validator, $field]);
+                }
+            }
+
             value = parseFloat(value);
-            return (options.inclusive === true || options.inclusive === undefined) ? (value <= options.value) : (value < options.value);
+            return (options.inclusive === true || options.inclusive === undefined)
+                    ? { valid: value <= compareTo, message: $.fn.bootstrapValidator.helpers.format($.fn.bootstrapValidator.i18n.lessThan['default'],   compareTo) }
+                    : { valid: value < compareTo,  message: $.fn.bootstrapValidator.helpers.format($.fn.bootstrapValidator.i18n.lessThan.notInclusive, compareTo) };
         }
     };
 }(window.jQuery));